+++ integration_tests/suite/test_phonebook.py
@@ -46,12 +46,57 @@ class TestPhonebookCRUD(BaseDirdIntegrationTest):
         result = self.post_phonebook('valid', valid_body)
         assert_that(result.status_code, equal_to(201))
 
+    def test_get_in_unknown_tenant(self):
+        tenants = {
+            'items': [
+                {
+                    'uuid': str(uuid4()),
+                    'name': 'valid',
+                }
+            ],
+            'total': 1,
+            'filtered': 1,
+        }
+        self.mock_auth_client.set_tenants(tenants)
+
+        valid_body = {'name': 'foobar'}
+        phonebook = self.post_phonebook('valid', valid_body).json()
+
+        tenants = {'items': [], 'total': 0, 'filtered': 0}
+        self.mock_auth_client.set_tenants(tenants)
+
+        result = self.get_phonebook('valid', phonebook['id'])
+        assert_that(result.status_code, equal_to(404))
+
+    def test_get_in_valid_tenant(self):
+        tenants = {
+            'items': [
+                {
+                    'uuid': str(uuid4()),
+                    'name': 'valid',
+                }
+            ],
+            'total': 1,
+            'filtered': 1,
+        }
+        self.mock_auth_client.set_tenants(tenants)
+
+        valid_body = {'name': 'foobar'}
+        phonebook = self.post_phonebook('valid', valid_body).json()
+
+        result = self.get_phonebook('valid', phonebook['id'])
+        assert_that(result.status_code, equal_to(200))
+        assert_that(result.json(), equal_to(phonebook))
+
     def test_all(self):
         tenant_1, tenant_2 = 'default', 'malicious'
         phonebook_1_body = {'name': 'integration',
                             'description': 'The integration test phonebook'}
         phonebook_1 = self.post_phonebook(tenant_1, phonebook_1_body).json()
-        assert_that(self.get_phonebook(tenant_1, phonebook_1['id']), equal_to(phonebook_1))
+        assert_that(
+            self.get_phonebook(tenant_1, phonebook_1['id']).json(),
+            equal_to(phonebook_1),
+        )
 
         expected = dict(phonebook_1_body)
         expected['id'] = ANY